argoslabs/shop/coupang/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# for rossum : https://rossum.ai/
# This rossum is out of date
requests
dpath
